[refactor] Move constructor class to init file
